@@ -1,12 +1,12 @@
 package com.entis.app.config.security;
 
-public final class SecurityConstants {
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
-    private SecurityConstants() {
-    }
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class SecurityConstants {
 
     public static final String AUTH_TOKEN_PREFIX = "Bearer ";
 
     public static final String AUTHORITIES_CLAIM = "authorities";
-
 }
