@@ -6,22 +6,22 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 import com.entis.app.config.security.SecurityConstants;
 import com.entis.app.config.security.properties.JWTProperties;
 import com.entis.app.entity.user.KnownAuthority;
+import com.entis.app.util.security.SecurityUtils;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.EnumSet;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.EnumSet;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
 
@@ -40,9 +40,7 @@ public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
                                     HttpServletResponse response,
                                     FilterChain chain) throws IOException, ServletException {
 
-        var securityContext = SecurityContextHolder.getContext();
-
-        var authentication = securityContext.getAuthentication();
+        Authentication authentication = SecurityUtils.getAuthentication();
         // if authenticated by other means, such as JWTAuthenticationFilter
         if (authentication != null && authentication.isAuthenticated()) {
             chain.doFilter(request, response);
@@ -59,7 +57,7 @@ public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
         String encodedJwt = header.substring(SecurityConstants.AUTH_TOKEN_PREFIX.length());
         authentication = getAuthentication(encodedJwt);
 
-        securityContext.setAuthentication(authentication);
+        SecurityUtils.setAuthentication(authentication);
         chain.doFilter(request, response);
     }
 
@@ -68,8 +66,8 @@ public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
         DecodedJWT decodedJWT;
         try {
             decodedJWT = JWT.require(algorithm)
-                    .build()
-                    .verify(encodedJwt);
+                .build()
+                .verify(encodedJwt);
         } catch (Exception e) {
             log.debug("Invalid JWT received", e);
             return null;
@@ -78,9 +76,9 @@ public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
         String email = decodedJWT.getSubject();
 
         Set<KnownAuthority> authorities = decodedJWT.getClaim(SecurityConstants.AUTHORITIES_CLAIM)
-                .asList(String.class).stream()
-                .map(KnownAuthority::valueOf)
-                .collect(Collectors.toCollection(() -> EnumSet.noneOf(KnownAuthority.class)));
+            .asList(String.class).stream()
+            .map(KnownAuthority::valueOf)
+            .collect(Collectors.toCollection(() -> EnumSet.noneOf(KnownAuthority.class)));
 
         return new UsernamePasswordAuthenticationToken(email, null, authorities);
     }
